Move to separate internal package
